backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-auth
